move dag round list
