+++ consensus/src/dag/dag_storage/mock.rs
@@ -1,9 +1,9 @@
 // Copyright © Aptos Foundation
 
 use std::any::Any;
-use crate::dag::dag::{DagInMem, DagInMem_Key, DagRoundList, WeakLinksCreator};
+use crate::dag::dag::{DagInMem, DagInMem_Key, WeakLinksCreator};
 use crate::dag::dag_storage::{DagStorage, DagStoreWriteBatch, ItemId};
-use crate::dag::types::MissingNodeIdToStatusMap;
+use crate::dag::types::{DagRoundList, MissingNodeIdToStatusMap};
 
 pub struct MockDagStoreWriteBatch {}
